+++ config/env/local.js
@@ -59,7 +59,12 @@ module.exports = {
 
     cronJobs: {
         enabled: true, // Enable/Disable cron jobs as a global setting
-        allowedHosts: [] // Define specific hosts (host names) to run cronjobs. Leave empty to start cron on any server running this application
+        allowedHosts: [], // Define specific hosts (host names) to run cronjobs. Leave empty to start cron on any server running this application
+        ensureIndexes: {
+            enabled: true,
+            startUp: true,
+            background: false
+        }
     },
 
 
@@ -88,24 +93,16 @@ module.exports = {
         mia: {
             url: 'mongodb://api:api@localhost:27017/mia',
             options: {
-                db: {
-                    w: 1 //write acknowledgement
-                },
-                server: {
-                    poolSize: 15
-                }
+                w: 1, //write acknowledgement
+                poolSize: 15
             }
         }
         // Tests db
         /*,tests: {
          url: 'mongodb://api:api@localhost:27017/tests',
          options: {
-         db: {
-         w: 1 //write acknowledgement
-         },
-         server: {
+         w: 1, //write acknowledgement
          poolSize: 15
-         }
          }
          }*/
     }